Prep tests for CVs
