+++ paths_cli/tests/wizard/test_openmm.py
@@ -7,34 +7,6 @@ from paths_cli.wizard.openmm import (
     _load_openmm_xml, _load_topology, openmm
 )
 
-@pytest.fixture
-def ad_openmm(tmpdir):
-    """
-    Provide directory with files to start alanine depeptide sim in OpenMM
-    """
-    mm = pytest.importorskip('simtk.openmm')
-    u = pytest.importorskip('simtk.unit')
-    openmmtools = pytest.importorskip('openmmtools')
-    md = pytest.importorskip('mdtraj')
-    testsystem = openmmtools.testsystems.AlanineDipeptideVacuum()
-    integrator = openmmtools.integrators.VVVRIntegrator(
-        300 * u.kelvin,
-        1.0 / u.picosecond,
-        2.0 * u.femtosecond
-    )
-    traj = md.Trajectory([testsystem.positions.value_in_unit(u.nanometer)],
-                         topology=testsystem.mdtraj_topology)
-    files = {'integrator.xml': integrator,
-             'system.xml': testsystem.system}
-    with tmpdir.as_cwd():
-        for fname, obj in files.items():
-            with open(fname, mode='w') as f:
-                f.write(mm.XmlSerializer.serialize(obj))
-
-        traj.save('ad.pdb')
-
-    return tmpdir
-
 
 @pytest.mark.parametrize('obj_type', ['system', 'integrator', 'foo'])
 def test_load_openmm_xml(ad_openmm, obj_type):
